pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "job-digest"
version = "0.1.0"
description = "Sources open job postings, ranks them against your profile using Claude, and publishes a Top 10 digest to GitHub Pages."
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "anthropic>=0.40",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "apify-client>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
job-digest = "job_digest.main:main"

[tool.ruff]
target-version = "py39"
line-length = 100
